#27 Added code to display the selected method.
More accurately, the display will show anything that is the source of
the double click, and displays the path through the tree to get to that
node. Once this is integrated with the method tokenizer, hopefully we'll
be able to parse that string into a method name, which will be used to
find the actual indexed code to display. If the method name is not found
(because the user double clicks a non-method), we can display an error,
or the timing.